requirements.txt
simple-term-menu
requests
tqdm
pyusb